can not delete store with accounts in it

+++ src/app/store/domain/store.service.ts
@@ -4,12 +4,16 @@ import { Store } from "./store.interface";
 import { StoreRepository } from "../infrastructure/store-repositor.type";
 import { SearchRequest } from "../../shared/base.request";
 import { ID } from "../../shared/abstract-repository/repository.interface";
+import { AccountService } from "src/app/account/domain/account.service";
 
 @Injectable()
 export class StoreService {
   constructor(
     @Inject("StoreRepository")
     private readonly storeRepo: StoreRepository,
+
+    @Inject(AccountService)
+    private readonly accountService: AccountService,
   ) {}
 
   private async verifyStoreName(name: string): Promise<void> {
@@ -50,6 +54,12 @@ export class StoreService {
 
   async removeStore(storeId: ID) {
     await this.findByIdOrFail(storeId);
+    const accounts = await this.accountService.getAccountsFromStore(storeId);
+    if (accounts.length > 0) {
+      throw new BadRequestException(
+        `Store with id ${storeId} has accounts associated with it`,
+      );
+    }
     await this.storeRepo.delete(storeId);
   }
